Clean up and reorganize unit tests
