+++ modules/core/src/test/java/org/locationtech/jts/algorithm/distance/DiscreteHausdorffDistanceTest.java
@@ -17,7 +17,7 @@ import org.locationtech.jts.io.ParseException;
 
 import junit.framework.TestCase;
 import junit.textui.TestRunner;
-import test.jts.junit.GeometryUtils;
+import test.jts.util.IOUtil;
 
 
 public class DiscreteHausdorffDistanceTest 
@@ -62,8 +62,8 @@ extends TestCase
   private void runTest(String wkt1, String wkt2, double expectedDistance) 
   throws ParseException
   {
-    Geometry g1 = GeometryUtils.readWKT(wkt1);
-    Geometry g2 = GeometryUtils.readWKT(wkt2);
+    Geometry g1 = IOUtil.readWKT(wkt1);
+    Geometry g2 = IOUtil.readWKT(wkt2);
     
     double distance = DiscreteHausdorffDistance.distance(g1, g2);
     assertEquals(distance, expectedDistance, TOLERANCE);
@@ -71,8 +71,8 @@ extends TestCase
   private void runTest(String wkt1, String wkt2, double densifyFrac, double expectedDistance) 
   throws ParseException
   {
-    Geometry g1 = GeometryUtils.readWKT(wkt1);
-    Geometry g2 = GeometryUtils.readWKT(wkt2);
+    Geometry g1 = IOUtil.readWKT(wkt1);
+    Geometry g2 = IOUtil.readWKT(wkt2);
     
     double distance = DiscreteHausdorffDistance.distance(g1, g2, densifyFrac);
     assertEquals(distance, expectedDistance, TOLERANCE);
