+++ modules/core/src/test/java/test/jts/perf/index/IndexTester.java
@@ -10,7 +10,7 @@
  *
  * http://www.eclipse.org/org/documents/edl-v10.php.
  */
-package test.jts.index;
+package test.jts.perf.index;
 
 import java.util.ArrayList;
 import java.util.Iterator;
