+++ modules/core/src/test/java/test/jts/util/IOUtil.java
@@ -12,8 +12,15 @@
 
 package test.jts.util;
 
+import java.io.IOException;
+import java.io.Reader;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.io.ParseException;
+import org.locationtech.jts.io.WKTFileReader;
 import org.locationtech.jts.io.WKTReader;
 
 public class IOUtil {
@@ -27,5 +34,39 @@ public class IOUtil {
 	      throw new RuntimeException(ex);
 	    }
 	  }
+
+    public static List readWKT(String[] inputWKT)
+    throws ParseException
+    {
+      ArrayList geometries = new ArrayList();
+      for (int i = 0; i < inputWKT.length; i++) {
+          geometries.add(IOUtil.reader.read(inputWKT[i]));
+      }
+      return geometries;
+    }
+
+    public static Geometry readWKT(String inputWKT)
+    throws ParseException
+    {
+    	return IOUtil.reader.read(inputWKT);
+    }
+
+    public static Collection readWKTFile(String filename) 
+    throws IOException, ParseException
+    {
+      WKTFileReader fileRdr = new WKTFileReader(filename, IOUtil.reader);
+      List geoms = fileRdr.read();
+      return geoms;
+    }
+
+    public static Collection readWKTFile(Reader rdr) 
+    throws IOException, ParseException
+    {
+      WKTFileReader fileRdr = new WKTFileReader(rdr, IOUtil.reader);
+      List geoms = fileRdr.read();
+      return geoms;
+    }
+
+    public static WKTReader reader = new WKTReader();
 
 }
